More updates to bulktester and cachepurging for Moodle 5.0.

+++ classes/task/cache_cleaner.php
@@ -20,7 +20,8 @@
  * This version doesn't do any authentication; it's assumed the server is
  * firewalled to accept connections only from Moodle.
  *
- * This class is used by the scheduler to cleanup the cache
+ * This class is used by the scheduler to cleanup the grading cache.
+ * Currently only cleans up the cache if it's in a file store.
  * Admins can change the schedule in Site Adminstration -> Server -> Scheduled Tasks -> Purge Old Coderunner Cache Entries
  *
  * @package    qtype_coderunner
@@ -59,6 +60,7 @@ class cache_cleaner extends \core\task\scheduled_task {
     /**
      * This task will purge old cache entries in file caches used
      * by the Coderunner grading cache definition.
+     * This task will only cleanup if a file store is being used.
      * Task does nothing if the grading cache isn't enabled
      * in the Coderunner admin settings. You might want
      * to purge the Coderunner grading cache manually
@@ -68,38 +70,46 @@ class cache_cleaner extends \core\task\scheduled_task {
         if (get_config('qtype_coderunner', 'enablegradecache')) {
             $purger = new cache_purger(usettl:true);
             $definition = $purger->get_coderunner_cache_definition();
-            $store = $purger->get_file_stores($definition);
-            $ttl = $purger->ttl;
-            if ($ttl) {
-                $days = round($ttl / 60 / 60 / 24, 4);
-                mtrace("Time to live (TTL) is $ttl seconds (= $days days)");
-            }
-            // $store->purge_old_entries();
-            // Use reflection to access the private cachestore_file method file_path_for_key
-            $reflection = new \ReflectionClass($store);
-            $filepathmethod = $reflection->getMethod('file_path_for_key');
-            $filepathmethod->setAccessible(true);
-
-            $keys = $store->find_all();
-            $originalcount = count($keys);
-            // Do a get on every key.
-            // The file cache get method should delete keys that are older than ttl but it doesn't...
-            $maxtime = cache::now() - $ttl;
-            foreach ($keys as $key) {
-                // Call the private method
-                $path = $filepathmethod->invoke($store, $key);
-                $filetime = filemtime($path);
-                if ($ttl && $filetime < $maxtime) {
-                    $store->delete($key);
+            $filestores = $purger->get_file_stores($definition);
+            if (count($filestores) > 0) {
+                $store = $purger->get_first_file_store($definition);
+                $ttl = $purger->ttl;
+                if ($ttl) {
+                    $days = round($ttl / 60 / 60 / 24, 4);
+                    mtrace("Time to live (TTL) is $ttl seconds (= $days days)");
                 }
-            }
+                // $store->purge_old_entries();
+                // Use reflection to access the private cachestore_file method file_path_for_key
+                $reflection = new \ReflectionClass($store);
+                $filepathmethod = $reflection->getMethod('file_path_for_key');
+                $filepathmethod->setAccessible(true);
+
+                $keys = $store->find_all();
+                $originalcount = count($keys);
+
+                // In theory doing a get on every key should delete old keys.
+                // The file cache get method should delete keys that are older than ttl but it doesn't...
+
+                $maxtime = cache::now() - $ttl;
+                foreach ($keys as $key) {
+                    // Call the private method.
+                    $path = $filepathmethod->invoke($store, $key);
+                    $filetime = filemtime($path);
+                    if ($ttl && $filetime < $maxtime) {
+                        $store->delete($key);
+                    }
+                }
                 // $value = $store->get($key);  // Would delete old key if fixed in file store.
-            $remainingkeys = $store->find_all();
-            $newcount = count($remainingkeys);
-            $purgedcount = $originalcount - $newcount;
-            mtrace("Originally found $originalcount keys.");
-            mtrace("$purgedcount keys purged.");
-            mtrace("$newcount keys were too young to die.");
+                $remainingkeys = $store->find_all();
+                $newcount = count($remainingkeys);
+                $purgedcount = $originalcount - $newcount;
+                mtrace("Originally found $originalcount keys.");
+                mtrace("$purgedcount keys purged.");
+                mtrace("$newcount keys were too young to die.");
+            } else {
+                mtrace("Grading cache not using a file store so nothing to do. You should probably unschedule this task.");
+                mtrace("If you're using a Redis store use the Redis purging task to manage cache size.");
+            }
         } else {
             mtrace("Grading cache not enabled so not purging.");
         }
